refactor: DB connector 분리 - DAO 추가로 인한 중복 발생
