+++ src/main/java/dao/DatabaseConnector.java
@@ -1,0 +1,7 @@
+package dao;
+
+import java.sql.Connection;
+
+public interface DatabaseConnector {
+    Connection getConnection();
+}
