+++ src/test/java/dao/JanggiBoardDaoTest.java
@@ -10,28 +10,13 @@ import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import java.sql.Connection;
-
 import static org.assertj.core.api.Assertions.assertThat;
 
 class JanggiBoardDaoTest {
-    private final JanggiBoardDao dao = new JanggiBoardDao();
+    private final JanggiBoardDao dao = new JanggiBoardDao(new JanggiDatabaseConnector());
     private final JanggiPosition position = JanggiPositionFactory.of(1, 1);
     private final JanggiTeam team = JanggiTeam.BLUE;
     private final Cannon piece = new Cannon(team);
-
-
-    @DisplayName("드라이버 연결 테스트")
-    @Test
-    void driverConnection() {
-        // given
-
-        // when
-        Connection connection = dao.getConnection();
-
-        // then
-        assertThat(connection).isNotNull();
-    }
 
     @Disabled
     @DisplayName("기물 추가 테스트")
